+++ engine/src/main/scala/cromwell/server/CromwellServer.scala
@@ -53,6 +53,8 @@ object CromwellServer {
 class CromwellServerActor(config: Config) extends CromwellRootActor with CromwellApiService with SwaggerService {
   implicit def executionContext = actorRefFactory.dispatcher
 
+  override val abortJobsOnTerminate = false
+
   override def actorRefFactory = context
   override def receive = handleTimeouts orElse runRoute(possibleRoutes)